Split the code into smaller more manageble files
