+++ css/main.css
@@ -1,0 +1,23 @@
+/* Main styles for DualShock Calibration GUI */
+
+dl.row dt {
+  font-weight: normal;
+}
+dl.row dd {
+  font-family: monospace;
+}
+
+#left-stick-card,
+#right-stick-card {
+  cursor: pointer;
+}
+
+.stick-card-active {
+  border: 1px solid #0d6efd !important;
+  box-shadow: 0 0 10px rgba(13, 110, 253, 0.3) !important;
+}
+
+.stick-card-active .card-header {
+  background-color: #0d6efd !important;
+  color: white !important;
+}
