+++ js/controller-manager.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import { sleep, la } from '../utils.js';
+import { sleep, la } from './utils.js';
 
 /**
 * Controller Manager - Manages the current controller instance and provides unified interface
@@ -8,7 +8,8 @@ import { sleep, la } from '../utils.js';
 class ControllerManager {
   constructor(uiDependencies = {}) {
     this.currentController = null;
-    this.l = uiDependencies.l || ((text) => text); // fallback to identity function
+    this.l = uiDependencies.l;
+    this.handleNvStatusUpdate = uiDependencies.handleNvStatusUpdate;
     this.has_changes_to_write = null; 
     this.inputHandler = null; // Callback function for input processing
 
@@ -66,6 +67,10 @@ class ControllerManager {
     return await this.currentController.getInfo();
   }
 
+  getFinetuneMaxValue() {
+    return this.currentController.getFinetuneMaxValue();
+  }
+
   /**
   * Set input report handler on the underlying device
   * @param {Function|null} handler Input report handler function or null to clear
@@ -79,7 +84,9 @@ class ControllerManager {
   * @returns {Promise<Object>} NVS status object
   */
   async queryNvStatus() {
-    return await this.currentController.queryNvStatus();
+    const nv = await this.currentController.queryNvStatus();
+    this.handleNvStatusUpdate(nv);
+    return nv;
   }
 
   /**
@@ -98,8 +105,8 @@ class ControllerManager {
     await this.currentController.writeFinetuneData(data);
   }
 
-  controllerType() {
-    return this.currentController.getType();
+  getModel() {
+    return this.currentController.getModel();
   }
 
   /**
@@ -130,21 +137,19 @@ class ControllerManager {
 
   /**
   * Update NVS changes status and UI
-  * @param {boolean} new_value Changes status
+  * @param {boolean} hasChanges Changes status
   */
-  setHasChangesToWrite(new_value) {
-    if (new_value === this.has_changes_to_write)
+  setHasChangesToWrite(hasChanges) {
+    if (hasChanges === this.has_changes_to_write)
       return;
 
-    if (new_value == true) {
-      $("#savechanges").prop("disabled", false);
-      $("#savechanges").addClass("btn-success").removeClass("btn-outline-secondary");
-    } else {
-      $("#savechanges").prop("disabled", true);
-      $("#savechanges").removeClass("btn-success").addClass("btn-outline-secondary");
-    }
+    const saveBtn = $("#savechanges");
+    saveBtn
+      .prop('disabled', !hasChanges)
+      .toggleClass('btn-success', hasChanges)
+      .toggleClass('btn-outline-secondary', !hasChanges);
 
-    this.has_changes_to_write = new_value;
+    this.has_changes_to_write = hasChanges;
   }
 
   // Unified controller operations that delegate to the current controller
@@ -170,6 +175,7 @@ class ControllerManager {
   */
   async nvsUnlock() {
     await this.currentController.nvsUnlock();
+    await this.queryNvStatus(); // Refresh NVS status
   }
 
   /**
@@ -181,6 +187,7 @@ class ControllerManager {
       throw new Error(this.l("NVS Lock failed: ") + String(res.error));
     }
 
+    await this.queryNvStatus(); // Refresh NVS status
     return res;
   }
 
@@ -193,7 +200,18 @@ class ControllerManager {
       const detail = res.code ? (this.l("Error ") + String(res.code)) : String(res.error || "");
       throw new Error(this.l("Stick calibration failed: ") + detail);
     }
-    return true;
+  }
+
+  /**
+  * Sample stick position during calibration
+  */
+  async calibrateSticksSample() {
+    const res = await this.currentController.calibrateSticksSample();
+    if (!res.ok) {
+      await sleep(500);
+      const detail = res.code ? (this.l("Error ") + String(res.code)) : String(res.error || "");
+      throw new Error(this.l("Stick calibration failed: ") + detail);
+    }
   }
 
   /**
@@ -208,20 +226,6 @@ class ControllerManager {
     }
 
     this.setHasChangesToWrite(true);
-    return true;
-  }
-
-  /**
-  * Sample stick position during calibration
-  */
-  async calibrateSticksSample() {
-    const res = await this.currentController.calibrateSticksSample();
-    if (!res.ok) {
-      await sleep(500);
-      const detail = res.code ? (this.l("Error ") + String(res.code)) : String(res.error || "");
-      throw new Error(this.l("Stick calibration failed: ") + detail);
-    }
-    return true;
   }
 
   /**
@@ -233,7 +237,6 @@ class ControllerManager {
       const detail = ret.code ? (this.l("Error ") + String(ret.code)) : String(ret.error || "");
       throw new Error(this.l("Range calibration failed: ") + detail);
     }
-    return true;
   }
 
   /**
@@ -270,22 +273,14 @@ class ControllerManager {
       la("multi_calibrate_sticks");
 
       progressCallback(20);
-
-      const okBegin = await this.calibrateSticksBegin();
-      if (!okBegin) {
-        return { success: false, message: this.l("Stick calibration failed to begin") };
-      }
-
+      await this.calibrateSticksBegin();
       progressCallback(30);
 
       // Sample multiple times during the process
       const sampleCount = 5;
       for (let i = 0; i < sampleCount; i++) {
         await sleep(100);
-        const okSample = await this.calibrateSticksSample();
-        if (!okSample) {
-          return { success: false, message: this.l("Stick calibration sampling failed") };
-        }
+        await this.calibrateSticksSample();
 
         // Progress from 30% to 80% during sampling
         const sampleProgress = 30 + ((i + 1) / sampleCount) * 50;
@@ -293,13 +288,9 @@ class ControllerManager {
       }
 
       progressCallback(90);
-
-      const okEnd = await this.calibrateSticksEnd();
-      if (!okEnd) {
-        return { success: false, message: this.l("Stick calibration failed to complete") };
-      }
-
+      await this.calibrateSticksEnd();
       progressCallback(100);
+
       return { success: true, message: this.l("Stick calibration completed") };
     } catch (e) {
       la("multi_calibrate_sticks_failed", {"r": e});
@@ -310,7 +301,7 @@ class ControllerManager {
   /**
   * Helper function to check if stick positions have changed
   */
-  sticksChanged(current, newValues) {
+  _sticksChanged(current, newValues) {
     return current.left.x !== newValues.left.x || current.left.y !== newValues.left.y ||
     current.right.x !== newValues.right.x || current.right.y !== newValues.right.y;
   }
@@ -319,7 +310,7 @@ class ControllerManager {
   * Generic button processing for DS4/DS5
   * Records button states and returns changes
   */
-  recordButtonStates(data, BUTTON_MAP, dpad_byte, l2_analog_byte, r2_analog_byte) {
+  _recordButtonStates(data, BUTTON_MAP, dpad_byte, l2_analog_byte, r2_analog_byte) {
     const changes = {};
 
     // Stick positions (always at bytes 0-3)
@@ -332,7 +323,7 @@ class ControllerManager {
       right: { x: new_rx, y: new_ry }
     };
 
-    if (this.sticksChanged(this.button_states.sticks, newSticks)) {
+    if (this._sticksChanged(this.button_states.sticks, newSticks)) {
       this.button_states.sticks = newSticks;
       changes.sticks = newSticks;
     }
@@ -390,22 +381,22 @@ class ControllerManager {
 
     const inputConfig = this.currentController.getInputConfig();
     const { buttonMap, dpadByte, l2AnalogByte, r2AnalogByte } = inputConfig;
-    const { touchpadOffset, batteryByte, isDS4 } = inputConfig;
+    const { touchpadOffset } = inputConfig;
 
     // Process button states using the device-specific configuration
-    const changes = this.recordButtonStates(data, buttonMap, dpadByte, l2AnalogByte, r2AnalogByte);
+    const changes = this._recordButtonStates(data, buttonMap, dpadByte, l2AnalogByte, r2AnalogByte);
 
     // Parse and store touch points if touchpad data is available
     if (touchpadOffset) {
-      this.touchPoints = this.parseTouchPoints(data, touchpadOffset);
+      this.touchPoints = this._parseTouchPoints(data, touchpadOffset);
     }
 
-    // Parse and store battery status if battery data is available
-    this.batteryStatus = this.parseBatteryStatus(data, batteryByte, isDS4);
+    // Parse and store battery status
+    this.batteryStatus = this._parseBatteryStatus(data);
 
     const result = {
       changes,
-      inputConfig: { buttonMap, isDS4 },
+      inputConfig: { buttonMap },
       touchPoints: this.touchPoints,
       batteryStatus: this.batteryStatus,
     };
@@ -419,7 +410,7 @@ class ControllerManager {
   * @param {number} offset - Offset to touchpad data
   * @returns {Array} Array of touch points with {active, id, x, y} properties
   */
-  parseTouchPoints(data, offset) {
+  _parseTouchPoints(data, offset) {
     // Returns array of up to 2 points: {active, id, x, y}
     const points = [];
     for (let i = 0; i < 2; i++) {
@@ -442,88 +433,21 @@ class ControllerManager {
 
   /**
   * Parse battery status from input data
-  * @param {DataView} data - Input data view
-  * @param {number} byte - Byte offset for battery data
-  * @param {boolean} isDS4 - Whether this is a DS4 controller
-  * @returns {Object} Battery status object with bat_txt, changed, bat_capacity, etc.
   */
-  parseBatteryStatus(data, byte, isDS4 = false) {
-    const bat = data.getUint8(byte);
-    let bat_capacity = 0, cable_connected = false, is_charging = false, is_error = false;
+  _parseBatteryStatus(data) {
+    const batteryInfo = this.currentController.parseBatteryStatus(data);
+    const bat_txt = this._batteryPercentToText(batteryInfo);
 
-    if (isDS4) {
-      // DS4: bat_data = low 4 bits, bat_status = bit 4
-      const bat_data = bat & 0x0f;
-      const bat_status = (bat >> 4) & 1;
-      if (bat_status == 1) {
-        cable_connected = true;
-        if (bat_data < 10) {
-          bat_capacity = Math.min(bat_data * 10 + 5, 100);
-          is_charging = true;
-        } else if (bat_data == 10) {
-          bat_capacity = 100;
-          is_charging = true;
-        } else if (bat_data == 11) {
-          bat_capacity = 100;
-          // charged
-        } else {
-          bat_capacity = 0;
-          is_error = true;
-        }
-      } else {
-        cable_connected = false;
-        if (bat_data < 10) {
-          bat_capacity = bat_data * 10 + 5;
-        } else {
-          bat_capacity = 100;
-        }
-      }
-    } else {
-      // DS5: bat_charge = low 4 bits, bat_status = high 4 bits
-      const bat_charge = bat & 0x0f;
-      const bat_status = bat >> 4;
-      if (bat_status == 0) {
-        bat_capacity = Math.min(bat_charge * 10 + 5, 100);
-      } else if (bat_status == 1) {
-        bat_capacity = Math.min(bat_charge * 10 + 5, 100);
-        is_charging = true;
-        cable_connected = true;
-      } else if (bat_status == 2) {
-        bat_capacity = 100;
-        cable_connected = true;
-      } else {
-        is_error = true;
-      }
-    }
-
-    // Generate battery text with icons
-    const bat_txt = this.batteryPercentToText(bat_capacity, is_charging, is_error);
-
-    // Check if battery text has changed
     const changed = bat_txt !== this._lastBatteryText;
     this._lastBatteryText = bat_txt;
 
-    // Update internal battery status
-    const batteryStatus = {
-      bat_txt,
-      changed,
-      bat_capacity,
-      cable_connected,
-      is_charging,
-      is_error
-    };
-
-    return batteryStatus;
+    return { bat_txt, changed, ...batteryInfo };
   }
 
   /**
   * Convert battery percentage to display text with icons
-  * @param {number} bat_charge - Battery charge percentage
-  * @param {boolean} is_charging - Whether battery is charging
-  * @param {boolean} is_error - Whether there's a battery error
-  * @returns {string} HTML string with battery status and icons
   */
-  batteryPercentToText(bat_charge, is_charging, is_error) {
+  _batteryPercentToText({bat_capacity, is_charging, is_error}) {
     if (is_error) {
       return '<font color="red">' + this.l("error") + '</font>';
     }
@@ -535,10 +459,10 @@ class ControllerManager {
       { threshold: 80, icon: 'fa-battery-three-quarters' },
     ];
 
-    const icon_txt = batteryIcons.find(item => bat_charge < item.threshold)?.icon || 'fa-battery-full';
-    const icon_full = '<i class="fa-solid ' + icon_txt + '"></i>';
+    const icon_txt = batteryIcons.find(item => bat_capacity < item.threshold)?.icon || 'fa-battery-full';
+    const icon_full = `<i class="fa-solid ${icon_txt}"></i>`;
     const bolt_txt = is_charging ? '<i class="fa-solid fa-bolt"></i>' : '';
-    return bat_charge + "%" + ' ' + bolt_txt + ' ' + icon_full;
+    return [`${bat_capacity}%`, icon_full, bolt_txt].join(' ');
   }
 
   /**
