+++ templates/calibrate-modal.html
@@ -1,0 +1,19 @@
+<!-- Modal -->
+<div class="modal fade" id="calibrateModal" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+  <div class="modal-dialog">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h1 class="modal-title fs-5" id="staticBackdropLabel">Calibrating center</h1>
+      </div>
+      <div class="modal-body">
+        <p class="ds-i18n">Recentering the controller sticks. </p>
+        <p class="ds-i18n">Please do not close this window and do not disconnect your controller. </p>
+        <div class="progress" role="progressbar" aria-label="Centering" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">
+            <div class="progress-bar" style="width: 0%"></div>
+        </div>
+      </div>
+      <div class="modal-footer">
+      </div>
+    </div>
+  </div>
+</div>
