+++ templates/edge-modal.html
@@ -1,0 +1,35 @@
+<div class="modal fade" id="edgeModal" tabindex="-1" aria-labelledby="modal-title" aria-hidden="true">
+  <div class="modal-dialog modal-xl modal-dialog-centered">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title ds-i18n">DualSense Edge Calibration</h5>
+        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+      </div>
+
+      <div class="modal-body p-4" id="donateBody">
+          <p class="ds-i18n">Support for calibrating DualSense Edge stick modules is now available as an <b>experimental feature</b>.</p>
+          <p class="ds-i18n">Please note: the stick modules on the DS Edge <b>cannot be calibrated via software alone</b>.</p>
+          <p class="ds-i18n">To store a custom calibration on the stick's internal memory, a <b>hardware modification</b> is required.</p>
+          <p class="ds-i18n">This involves temporarily disabling write protection by applying <b>+1.8V</b> to a specific test point on each module.</p>
+          <p></p>
+
+          <p><span class="ds-i18n">You can do this in two ways:</span>
+          <ul>
+            <li class="ds-i18n"><b>Internally</b>: by soldering a wire from a +1.8V source to the write-protect TP.</li>
+            <li class="ds-i18n"><b>Externally</b>: by applying +1.8V directly to the visible test point without opening the controller.</li>
+          </ul>
+        </p>
+
+        <p><b><span class="ds-i18n">This is only for advanced users. If you're not sure what you're doing, please do not attempt it.</span></b></p>
+        <p><span class="ds-i18n">More details and images</span>&nbsp;<a href="https://github.com/lewy20041/Dualsense_Edge_Modules_Callibration">here</a>.</p>
+
+        <p class="ds-i18n">We are not responsible for any damage caused by attempting this modification.</p>
+        <p class="ds-i18n">For more info or help, feel free to reach out on Discord.</p>
+
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-primary ds-i18n" data-bs-dismiss="modal">Understood</button>
+      </div>
+    </div>
+  </div>
+</div>
