+++ templates/edge-progress-modal.html
@@ -1,0 +1,20 @@
+<!-- Edge in progress Modal -->
+<div class="modal fade" id="edgeProgressModal" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="edgeProgressLabel" aria-hidden="true">
+  <div class="modal-dialog  modal-dialog-centered">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h1 class="modal-title fs-5 ds-i18n" id="edgeProgressLabel">Storing calibration...</h1>
+      </div>
+      <div class="modal-body">
+        <p class="ds-i18n">Calibration is being stored in the stick modules.</p>
+        <p class="ds-i18n">Please do not close this window and do not disconnect your controller. </p>
+
+        <div class="progress" role="progressbar" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">
+        <div id="dsedge-progress" class="progress-bar progress-bar-striped progress-bar-animated" style="width: 0%"></div>
+        </div>
+
+
+      </div>
+    </div>
+  </div>
+</div>
