+++ templates/range-modal.html
@@ -1,0 +1,17 @@
+<!-- Modal -->
+<div class="modal fade" id="rangeModal" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+  <div class="modal-dialog">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h1 class="modal-title fs-5 ds-i18n" id="staticBackdropLabel">Range calibration</h1>
+      </div>
+      <div class="modal-body">
+        <p class="ds-i18n"><b>The controller is now sampling data!</b></p>
+        <p class="ds-i18n">Rotate the sticks slowly to cover the whole range. Press "Done" when completed.</p>
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-primary ds-i18n" onclick="calibrate_range_on_close()">Done</button>
+      </div>
+    </div>
+  </div>
+</div>
