+++ templates/welcome-modal.html
@@ -1,0 +1,23 @@
+<!-- Welcome Modal -->
+<div class="modal fade" id="welcomeModal" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="welcomeModalLabel" aria-hidden="true">
+  <div class="modal-dialog modal-dialog-centered modal-lg">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h1 class="modal-title fs-5 ds-i18n" id="welcomeModalLabel">Welcome to the Calibration GUI</h1>
+      </div>
+      <div class="modal-body">
+      <p class="ds-i18n">Just few things to know before you can start:</p>
+      <ul>
+        <li class="ds-i18n">This website is not affiliated with Sony, PlayStation &amp; co.</li>
+        <li class="ds-i18n">This service is provided without warranty. Use at your own risk.</li>
+        <li class="ds-i18n">This website uses analytics to improve the service.</li>
+        <li class="ds-i18n">Keep the internal battery of the controller connected and ensure it is well charged. If the battery dies during operations, the controller will be damaged and rendered unusable.</li>
+        <li class="ds-i18n">Before doing the permanent calibration, try the temporary one to ensure that everything is working well.</li>
+      </ul>
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-primary ds-i18n" onclick="welcome_accepted();">Understood</button>
+      </div>
+    </div>
+  </div>
+</div>
